Move the Map view of NamedNodeMap to its companion object.
So that it is always available.

We also make the implementation class private.
