+++ src/main/scala/org/scalajs/dom/lib.scala
@@ -10,6 +10,8 @@
 package org.scalajs.dom
 
 import org.scalajs.dom.experimental.ReadableStream
+import scala.collection.mutable
+import scala.language.implicitConversions
 import scala.scalajs.js
 import scala.scalajs.js.annotation._
 import scala.scalajs.js.typedarray.ArrayBuffer
@@ -6177,6 +6179,12 @@ class NamedNodeMap extends js.Object {
   def getNamedItemNS(namespaceURI: String, localName: String): Attr = js.native
 
   def setNamedItemNS(arg: Attr): Attr = js.native
+}
+
+object NamedNodeMap {
+  implicit def namedNodeMapAsMap(
+      namedNodeMap: NamedNodeMap): mutable.Map[String, Attr] =
+    new NamedNodeMapMap(namedNodeMap)
 }
 
 @js.native
